refactor:yaml related functions
