+++ pkg/models/http.go
@@ -4,6 +4,17 @@ import "time"
 
 type Method string
 
+type HttpSchema struct {
+	Metadata         map[string]string      `json:"metadata" yaml:"metadata"`
+	Request          HttpReq                `json:"req" yaml:"req"`
+	Response         HttpResp               `json:"resp" yaml:"resp"`
+	Objects          []*OutputBinary        `json:"objects" yaml:"objects"`
+	Assertions       map[string]interface{} `json:"assertions" yaml:"assertions,omitempty"`
+	Created          int64                  `json:"created" yaml:"created,omitempty"`
+	ReqTimestampMock time.Time              `json:"reqTimestampMock" yaml:"reqTimestampMock,omitempty"`
+	ResTimestampMock time.Time              `json:"resTimestampMock" yaml:"resTimestampMock,omitempty"`
+}
+
 type HttpReq struct {
 	Method     Method            `json:"method" yaml:"method"`
 	ProtoMajor int               `json:"proto_major" yaml:"proto_major"` // e.g. 1
