--- pkg/platform/yaml/spec/grpc.go
@@ -1,14 +1,0 @@
-package spec
-
-import (
-	"time"
-
-	"go.keploy.io/server/pkg/models"
-)
-
-type GrpcSpec struct {
-	GrpcReq          models.GrpcReq  `json:"grpcReq" yaml:"grpcReq"`
-	GrpcResp         models.GrpcResp `json:"grpcResp" yaml:"grpcResp"`
-	ReqTimestampMock time.Time       `json:"reqTimestampMock" yaml:"reqTimestampMock,omitempty"`
-	ResTimestampMock time.Time       `json:"resTimestampMock" yaml:"resTimestampMock,omitempty"`
-}
