+++ pkg/platform/yaml/utils.go
@@ -1,304 +1,1 @@
 package yaml
-
-import (
-	"encoding/json"
-	"fmt"
-	"io/fs"
-	"net/http"
-	"os"
-	"path/filepath"
-	"reflect"
-
-	"strconv"
-	"strings"
-
-	"go.keploy.io/server/pkg/models"
-	"go.uber.org/zap"
-)
-
-func FlattenHttpResponse(h http.Header, body string) (map[string][]string, error) {
-	m := map[string][]string{}
-	for k, v := range h {
-		m["header."+k] = []string{strings.Join(v, "")}
-	}
-	err := AddHttpBodyToMap(body, m)
-	if err != nil {
-		return m, err
-	}
-	return m, nil
-}
-
-// Flatten takes a map and returns a new one where nested maps are replaced
-// by dot-delimited keys.
-// examples of valid jsons - https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/JSON/parse#examples
-func Flatten(j interface{}) map[string][]string {
-	if j == nil {
-		return map[string][]string{"": {""}}
-	}
-	o := make(map[string][]string)
-	x := reflect.ValueOf(j)
-	switch x.Kind() {
-	case reflect.Map:
-		m, ok := j.(map[string]interface{})
-		if !ok {
-			return map[string][]string{}
-		}
-		for k, v := range m {
-			nm := Flatten(v)
-			for nk, nv := range nm {
-				fk := k
-				if nk != "" {
-					fk = fk + "." + nk
-				}
-				o[fk] = nv
-			}
-		}
-	case reflect.Bool:
-		o[""] = []string{strconv.FormatBool(x.Bool())}
-	case reflect.Float64:
-		o[""] = []string{strconv.FormatFloat(x.Float(), 'E', -1, 64)}
-	case reflect.String:
-		o[""] = []string{x.String()}
-	case reflect.Slice:
-		child, ok := j.([]interface{})
-		if !ok {
-			return map[string][]string{}
-		}
-		for _, av := range child {
-			nm := Flatten(av)
-			for nk, nv := range nm {
-				if ov, exists := o[nk]; exists {
-					o[nk] = append(ov, nv...)
-				} else {
-					o[nk] = nv
-				}
-			}
-		}
-	default:
-		fmt.Println(Emoji, "found invalid value in json", j, x.Kind())
-	}
-	return o
-}
-
-func AddHttpBodyToMap(body string, m map[string][]string) error {
-	// add body
-	if json.Valid([]byte(body)) {
-		var result interface{}
-
-		err := json.Unmarshal([]byte(body), &result)
-		if err != nil {
-			return err
-		}
-		j := Flatten(result)
-		for k, v := range j {
-			nk := "body"
-			if k != "" {
-				nk = nk + "." + k
-			}
-			m[nk] = v
-		}
-	} else {
-		// add it as raw text
-		m["body"] = []string{body}
-	}
-	return nil
-}
-
-func FindNoisyFields(m map[string][]string, comparator func(string, []string) bool) []string {
-	var noise []string
-	for k, v := range m {
-		if comparator(k, v) {
-			noise = append(noise, k)
-		}
-	}
-	return noise
-}
-
-func CompareHeaders(h1 http.Header, h2 http.Header, res *[]models.HeaderResult, noise map[string]string) bool {
-	if res == nil {
-		return false
-	}
-	match := true
-	_, isHeaderNoisy := noise["header"]
-	for k, v := range h1 {
-		_, isNoisy := noise[k]
-		isNoisy = isNoisy || isHeaderNoisy
-		val, ok := h2[k]
-		if !isNoisy {
-			if !ok {
-				if checkKey(res, k) {
-					*res = append(*res, models.HeaderResult{
-						Normal: false,
-						Expected: models.Header{
-							Key:   k,
-							Value: v,
-						},
-						Actual: models.Header{
-							Key:   k,
-							Value: nil,
-						},
-					})
-				}
-
-				match = false
-				continue
-			}
-			if len(v) != len(val) {
-				if checkKey(res, k) {
-					*res = append(*res, models.HeaderResult{
-						Normal: false,
-						Expected: models.Header{
-							Key:   k,
-							Value: v,
-						},
-						Actual: models.Header{
-							Key:   k,
-							Value: val,
-						},
-					})
-				}
-				match = false
-				continue
-			}
-			for i, e := range v {
-				if val[i] != e {
-					if checkKey(res, k) {
-						*res = append(*res, models.HeaderResult{
-							Normal: false,
-							Expected: models.Header{
-								Key:   k,
-								Value: v,
-							},
-							Actual: models.Header{
-								Key:   k,
-								Value: val,
-							},
-						})
-					}
-					match = false
-					continue
-				}
-			}
-		}
-		if checkKey(res, k) {
-			*res = append(*res, models.HeaderResult{
-				Normal: true,
-				Expected: models.Header{
-					Key:   k,
-					Value: v,
-				},
-				Actual: models.Header{
-					Key:   k,
-					Value: val,
-				},
-			})
-		}
-	}
-	for k, v := range h2 {
-		_, isNoisy := noise[k]
-		isNoisy = isNoisy || isHeaderNoisy
-		val, ok := h1[k]
-		if isNoisy && checkKey(res, k) {
-			*res = append(*res, models.HeaderResult{
-				Normal: true,
-				Expected: models.Header{
-					Key:   k,
-					Value: val,
-				},
-				Actual: models.Header{
-					Key:   k,
-					Value: v,
-				},
-			})
-			continue
-		}
-		if !ok {
-			if checkKey(res, k) {
-				*res = append(*res, models.HeaderResult{
-					Normal: false,
-					Expected: models.Header{
-						Key:   k,
-						Value: nil,
-					},
-					Actual: models.Header{
-						Key:   k,
-						Value: v,
-					},
-				})
-			}
-
-			match = false
-		}
-	}
-	return match
-}
-
-func checkKey(res *[]models.HeaderResult, key string) bool {
-	for _, v := range *res {
-		if key == v.Expected.Key {
-			return false
-		}
-	}
-	return true
-}
-
-func Contains(elems []string, v string) bool {
-	for _, s := range elems {
-		if v == s {
-			return true
-		}
-	}
-	return false
-}
-
-func NewSessionIndex(path string, Logger *zap.Logger) (string, error) {
-	indx := 0
-	dir, err := os.OpenFile(path, os.O_RDONLY, fs.FileMode(os.O_RDONLY))
-	if err != nil {
-		Logger.Debug("creating a folder for the keploy generated testcases", zap.Error(err))
-		return fmt.Sprintf("%s%v", models.TestSetPattern, indx), nil
-	}
-
-	files, err := dir.ReadDir(0)
-	if err != nil {
-		return "", err
-	}
-
-	for _, v := range files {
-		// fmt.Println("name for the file", v.Name())
-		fileName := filepath.Base(v.Name())
-		fileNamePackets := strings.Split(fileName, "-")
-		if len(fileNamePackets) == 3 {
-			fileIndx, err := strconv.Atoi(fileNamePackets[2])
-			if err != nil {
-				Logger.Debug("failed to convert the index string to integer", zap.Error(err))
-				continue
-			}
-			if indx < fileIndx+1 {
-				indx = fileIndx + 1
-			}
-		}
-	}
-	return fmt.Sprintf("%s%v", models.TestSetPattern, indx), nil
-}
-
-func ReadSessionIndices(path string, Logger *zap.Logger) ([]string, error) {
-	indices := []string{}
-	dir, err := os.OpenFile(path, os.O_RDONLY, fs.FileMode(os.O_RDONLY))
-	if err != nil {
-		Logger.Debug("creating a folder for the keploy generated testcases", zap.Error(err))
-		return indices, nil
-	}
-
-	files, err := dir.ReadDir(0)
-	if err != nil {
-		return indices, err
-	}
-
-	for _, v := range files {
-		if v.Name() != "testReports" {
-			indices = append(indices, v.Name())
-		}
-	}
-	return indices, nil
-}
