+++ pkg/platform/yaml/yamlEncoders/util.go
@@ -1,0 +1,110 @@
+package yamlencoders
+
+import (
+	"encoding/json"
+	"fmt"
+	"net/http"
+	"reflect"
+	"strconv"
+	"strings"
+
+	"go.keploy.io/server/utils"
+)
+
+func FlattenHttpResponse(h http.Header, body string) (map[string][]string, error) {
+	m := map[string][]string{}
+	for k, v := range h {
+		m["header."+k] = []string{strings.Join(v, "")}
+	}
+	err := AddHttpBodyToMap(body, m)
+	if err != nil {
+		return m, err
+	}
+	return m, nil
+}
+
+func AddHttpBodyToMap(body string, m map[string][]string) error {
+	// add body
+	if json.Valid([]byte(body)) {
+		var result interface{}
+
+		err := json.Unmarshal([]byte(body), &result)
+		if err != nil {
+			return err
+		}
+		j := Flatten(result)
+		for k, v := range j {
+			nk := "body"
+			if k != "" {
+				nk = nk + "." + k
+			}
+			m[nk] = v
+		}
+	} else {
+		// add it as raw text
+		m["body"] = []string{body}
+	}
+	return nil
+}
+
+// Flatten takes a map and returns a new one where nested maps are replaced
+// by dot-delimited keys.
+// examples of valid jsons - https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/JSON/parse#examples
+func Flatten(j interface{}) map[string][]string {
+	if j == nil {
+		return map[string][]string{"": {""}}
+	}
+	o := make(map[string][]string)
+	x := reflect.ValueOf(j)
+	switch x.Kind() {
+	case reflect.Map:
+		m, ok := j.(map[string]interface{})
+		if !ok {
+			return map[string][]string{}
+		}
+		for k, v := range m {
+			nm := Flatten(v)
+			for nk, nv := range nm {
+				fk := k
+				if nk != "" {
+					fk = fk + "." + nk
+				}
+				o[fk] = nv
+			}
+		}
+	case reflect.Bool:
+		o[""] = []string{strconv.FormatBool(x.Bool())}
+	case reflect.Float64:
+		o[""] = []string{strconv.FormatFloat(x.Float(), 'E', -1, 64)}
+	case reflect.String:
+		o[""] = []string{x.String()}
+	case reflect.Slice:
+		child, ok := j.([]interface{})
+		if !ok {
+			return map[string][]string{}
+		}
+		for _, av := range child {
+			nm := Flatten(av)
+			for nk, nv := range nm {
+				if ov, exists := o[nk]; exists {
+					o[nk] = append(ov, nv...)
+				} else {
+					o[nk] = nv
+				}
+			}
+		}
+	default:
+		fmt.Println(utils.Emoji, "found invalid value in json", j, x.Kind())
+	}
+	return o
+}
+
+func FindNoisyFields(m map[string][]string, comparator func(string, []string) bool) []string {
+	var noise []string
+	for k, v := range m {
+		if comparator(k, v) {
+			noise = append(noise, k)
+		}
+	}
+	return noise
+}
